LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-link-element/stylesheet-media-change-no-fetch-expected.txt

hello


FAIL stylesheet should not re-fetch when media changes. The media change should apply immediately assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
